refact(task state enums): move to enums folder
https://app.asana.com/0/287748850352677/290559930627325
