+++ src/app/pomodoro/tasks.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {Store} from '@ngrx/store';
 import {Observable} from 'rxjs';
-import {TaskState} from './task-state.enum';
+import {TaskState} from '../_enums/task-state.enum';
 import {ObjectState} from '../_enums/object-state.enum';
 import {Task} from '../_domains/task';
 import {Tasks} from '../_reducers/tasks.reducer';
